fixed placement of dropdown-contents in navbar
Now, if the dropdowns would reach outside of the window, they move neatly towards the other side.

Also, the white-spaces between two adjacent dropbuttons was removed, such that font-size:0 could be removed from #navbar.
Now it should be easier to use the unit em.

TODO: Cleaning up.

TO CHECK: Possible bug in other browsers:
In navbar.html we listen to the mouseenter-event for all dropbuttons (in order to adjust placement of the dropdown-content); here we silently assume that CSS already ran the update for dropbtn:hover. I do not know whether this order is the same for all browsers. For firefox and chrome on linux it works.

	modified:   lmfdb/templates/navbar.html
	modified:   lmfdb/templates/style.css