added css_grid styling for search boxes and spacers in forms
That is much preferred over colspan and rowspan options from before.
For the result see for example:
 - /Motive/Hypergeometric/Q/
 - /NumberField/
 - /EllipticCurve/

The biggest challenge was (and still is) abvar, due to pairs of search boxes that are related.
TODO: The currently best way to solve this is to allow to group them together (-> new container class in search_boxes.py) such that they are shown together, either side by side or better: above each other.

TODO: Gridlayout for the buttons!

TODO: Clean up search_boxes.py and everything that depends on it (no more colspans etc...).

	modified:   lmfdb/abvar/fq/main.py
	modified:   lmfdb/bianchi_modular_forms/bianchi_modular_form.py
	modified:   lmfdb/ecnf/main.py
	modified:   lmfdb/genus2_curves/main.py
	modified:   lmfdb/hypergm/main.py
	modified:   lmfdb/number_fields/number_field.py
	modified:   lmfdb/templates/style.css
	modified:   lmfdb/utils/__init__.py
	modified:   lmfdb/utils/search_boxes.py
	modified:   lmfdb/utils/utilities.py